Revisado

+++ practica5_parte3_lv.sh
@@ -31,20 +31,22 @@ do
 
 	else	
 		exite=$(sudo lvscan | grep "$ruta" )
-		echo "$exite"
 		if [  -z "$exite"  ] #Si no hay contenido en la variables es que no existe
 		then #Tenemos que crear el volumen logico
 			sudo lvcreate -L "$tamanyo" --name "$volumenlogico" "$grupovolumen"
-			sudo mkfs -t "$tsistemaficheros" "$ruta"
-			if [ ! -d "$directorio" ] #Comprobamos si existe el directorio
-			then	
-				echo "Creando directorio: $directorio"
-				sudo mkdir "$directorio"
-			fi #Montamos el sistema de ficheros
-			sudo mount -t  "$tsistemaficheros" "$ruta" "$directorio"
-			ID=$(sudo blkid "$ruta" | cut -d ' ' -f2 | cut -d '=' -f2  )
-			#Añadimos el volumen lógico para su correcto montaje en el arranque
-			sudo su -c "echo "$ID $directorio $tsistemaficheros errors=remount 0 2" >> /etc/fstab"
+			
+			if [ $? -eq 0 ]; then
+				sudo mkfs -t "$tsistemaficheros" "$ruta"
+				if [ ! -d "$directorio" ] #Comprobamos si existe el directorio
+				then	
+					echo "Creando directorio: $directorio"
+					sudo mkdir -p "$directorio"
+				fi #Montamos el sistema de ficheros
+				sudo mount -t  "$tsistemaficheros" "$ruta" "$directorio"
+				ID=$(sudo blkid "$ruta" | cut -d ' ' -f2  )
+				#Añadimos el volumen lógico para su correcto montaje en el arranque
+				sudo su -c "echo \"$ID $directorio $tsistemaficheros errors=remount 0 2\" >> /etc/fstab"
+			fi
 		else #El volúmen lógico ya existía y lo extendemos
 			sudo lvextend -L "$tamanyo" "$ruta"
 			sudo resize2fs "$ruta"
